[REFACTOR] 7. Factor out PostPageWrapper
